easy-rpc-spring-boot-starter/src/main/java/org/example/rpc/client/network/netty/NettyRpcClient.java
package org.example.rpc.client.network.netty;

import io.netty.bootstrap.Bootstrap;
import io.netty.channel.ChannelInitializer;
import io.netty.channel.ChannelOption;
import io.netty.channel.ChannelPipeline;
import io.netty.channel.nio.NioEventLoopGroup;
import io.netty.channel.socket.SocketChannel;
import io.netty.channel.socket.nio.NioSocketChannel;
import org.example.rpc.client.network.RpcClient;
import org.example.rpc.common.ServiceInterfaceInfo;

/**
 * Netty实现客户端消息发送
 *
 * @Author Roc
 * @Date 2024/11/2 15:15
 */
public class NettyRpcClient implements RpcClient {
    @Override
    public byte[] sendMessage(byte[] data, ServiceInterfaceInfo serviceInterfaceInfo) throws Exception {
        final String ip = serviceInterfaceInfo.getIp();
        final Integer port = serviceInterfaceInfo.getPort();

        ClientChannelHandler handler = new ClientChannelHandler(data);
        final Bootstrap bootstrap = new Bootstrap()
                .group(new NioEventLoopGroup())
                .channel(NioSocketChannel.class)
                .option(ChannelOption.TCP_NODELAY, true)
                .handler(new ChannelInitializer<SocketChannel>() {
                    @Override
                    protected void initChannel(SocketChannel socketChannel) throws Exception {
                        final ChannelPipeline pipeline = socketChannel.pipeline();
                        pipeline.addLast(handler);
                    }
                });
        bootstrap.connect(ip, port).sync();
        return handler.response();
    }
}
